+++ src/render/primitives/PacketStateHelpers.h
@@ -1,11 +1,9 @@
 #pragma once
-#include "core/SimdFeatures.h"
 #include "render/State.h"
 #include "render/primitives/Primitive.h"
 
 #include <string>
 
-#if RAYTRACER_SIMD_SSE || RAYTRACER_SIMD_NEON
 namespace render {
   inline void packetHit(State& state, const Primitive* primitive, const std::string& reason) {
     if (state.traceEvents) {
@@ -23,4 +21,3 @@ namespace render {
     }
   }
 }
-#endif